Add classes
First try at it. Have not tested.
