+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@
 
 -------------------------------------------------------------------
 
-Hello! I'm Skibidi!
+Hello! I'm Ui!
 What can I do for you?
 
 -------------------------------------------------------------------
